user auth fixed

+++ app/Http/Controllers/TasksController.php
@@ -37,11 +37,7 @@ class TasksController extends Controller
      */
     public function create()
     {
-        if (\Auth::check()) {
-            return view('tasks.create');
-        }else{
-            return redirect('/');
-        }
+        return view('tasks.create');
     }
 
     /**
@@ -116,10 +112,17 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        return redirect('/');
+        
+        if (\Auth::id() == $task->user_id)
+        {
+            $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
+            return redirect('/');
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -131,7 +134,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
-        return redirect('/');
+        
+        if (\Auth::id() == $task->user_id)
+        {
+            $task->delete();
+            return redirect('/');
+        }else{
+            return redirect('/');
+        }
     }
 }
